Reinstate original table names
Notes:

Had to reinstate the original table names due to some complications
that will come up when changing them in our case. Any changes to the
names (at least for now, during the module separation) could prevent
a user from using Magento's setup:upgrade command. This happens because
the user has tables left in the database that will have references to
the old names. If we want to rename a table we will have to make that
decision at a later date.

M2C-20